refactor: move User domain model to the domain.model package
