+++ src/main/java/br/com/aegro/datasync/user/domain/UserRepository.java
@@ -1,5 +1,7 @@
 package br.com.aegro.datasync.user.domain;
 
+import br.com.aegro.datasync.user.domain.model.User;
+
 import java.util.Optional;
 
 public interface UserRepository {
