+++ src/main/java/br/com/aegro/datasync/user/persistence/JpaUserRepository.java
@@ -1,6 +1,6 @@
 package br.com.aegro.datasync.user.persistence;
 
-import br.com.aegro.datasync.user.domain.User;
+import br.com.aegro.datasync.user.domain.model.User;
 import br.com.aegro.datasync.user.domain.UserRepository;
 import br.com.aegro.datasync.user.persistence.dao.UserDao;
 import br.com.aegro.datasync.user.persistence.mapper.UserMapper;
