+++ src/main/java/br/com/aegro/datasync/user/persistence/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package br.com.aegro.datasync.user.persistence.mapper;
 
-import br.com.aegro.datasync.user.domain.User;
+import br.com.aegro.datasync.user.domain.model.User;
 import br.com.aegro.datasync.user.persistence.model.UserData;
 
 public interface UserMapper {
